refactor: Add ICombatant interface to unify player and mob combat
- Create ICombatant interface with combat properties and skill methods
- Implement ICombatant in PlayerState and MobInstance
- Update ISkillHandler to use ICombatant instead of PlayerState
- Refactor KickSkill to work with any ICombatant (players and mobs)
- Replace separate KickPlayer/KickMob methods with unified ExecuteKick
- Fix int->short conversions for MobInstance.HitPoints

This allows:
- Skills to work for both players and mobs
- Single code path for PvP, PvE, and future MvM combat
- Zero duplicate combat logic in skill handlers
- Extensibility for future combatant types (pets, summons, NPCs)

+++ src/EliteMud.Server/Adapters/Commands/Kill/KillCommandHandler.cs
@@ -174,7 +174,7 @@ internal sealed class KillCommandHandler : ICommandHandler
         // Perform initial attack
         int mobMaxHp = mob.Definition.MaxHitPoints;
         int damage = CombatCalculator.CalculateBareDamage(attacker.Player);
-        mob.HitPoints -= damage;
+        mob.HitPoints -= (short)damage;
         
         // Format legacy combat messages
         var attackerMsg = CombatMessageFormatter.FormatCombatMessage(
